# lib/api-zod/src/generated/types/application.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FSV Capital - Startup Funding Application API
 * OpenAPI spec version: 0.1.0
 */
import type { CoreTeamMember } from './coreTeamMember';
import type { ScoreBreakdown } from './scoreBreakdown';

export interface Application {
  id: string;
  applicationId: string;
  submittedAt: string;
  status: string;
  startupName: string;
  /** @nullable */
  websiteUrl?: string | null;
  founderNames?: string;
  contactEmail: string;
  /** @nullable */
  contactNumber?: string | null;
  /** @nullable */
  linkedinFounder?: string | null;
  /** @nullable */
  linkedinCompany?: string | null;
  /** @nullable */
  headquarters?: string | null;
  /** @nullable */
  yearOfIncorporation?: number | null;
  /** @nullable */
  problemStatement?: string | null;
  /** @nullable */
  solutionOverview?: string | null;
  sector?: string;
  /** @nullable */
  businessModel?: string | null;
  stage?: string;
  /** @nullable */
  coreProduct?: string | null;
  techStack?: string[];
  /** @nullable */
  usp?: string | null;
  /** @nullable */
  hasPatents?: boolean | null;
  /** @nullable */
  patentDetails?: string | null;
  /** @nullable */
  demoLink?: string | null;
  /** @nullable */
  tam?: string | null;
  /** @nullable */
  sam?: string | null;
  /** @nullable */
  som?: string | null;
  /** @nullable */
  customerSegment?: string | null;
  /** @nullable */
  competitors?: string | null;
  /** @nullable */
  competitiveAdvantage?: string | null;
  /** @nullable */
  monthlyRevenue?: number | null;
  /** @nullable */
  annualRevenue?: number | null;
  /** @nullable */
  growthRate?: number | null;
  /** @nullable */
  customerCount?: number | null;
  /** @nullable */
  partnerships?: string | null;
  /** @nullable */
  achievements?: string | null;
  /** @nullable */
  fundingRaised?: number | null;
  /** @nullable */
  previousInvestors?: string | null;
  /** @nullable */
  burnRate?: number | null;
  /** @nullable */
  runway?: number | null;
  /** @nullable */
  projectionY1?: number | null;
  /** @nullable */
  projectionY2?: number | null;
  /** @nullable */
  projectionY3?: number | null;
  /** @nullable */
  amountRaising?: number | null;
  /** @nullable */
  currency?: string | null;
  /** @nullable */
  fundingStage?: string | null;
  /** @nullable */
  equityOffered?: number | null;
  useOfFunds?: string[];
  /** @nullable */
  useOfFundsDetail?: string | null;
  /** @nullable */
  founderEducation?: string | null;
  /** @nullable */
  founderExperience?: string | null;
  /** @nullable */
  teamSize?: number | null;
  coreTeam?: CoreTeamMember[];
  /** @nullable */
  advisors?: string | null;
  /** @nullable */
  whyFSV?: string | null;
  /** @nullable */
  fsvValueAdd?: string | null;
  /** @nullable */
  openToMentorship?: boolean | null;
  /** @nullable */
  isRegistered?: boolean | null;
  /** @nullable */
  hasLegalIssues?: boolean | null;
  /** @nullable */
  legalExplanation?: string | null;
  /** @nullable */
  consentGiven?: boolean | null;
  /** @nullable */
  dealScore?: number | null;
  scoreBreakdown?: ScoreBreakdown;
  /** @nullable */
  aiAnalysis?: string | null;
}