# lib/api-zod/src/generated/types/applicationInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FSV Capital - Startup Funding Application API
 * OpenAPI spec version: 0.1.0
 */
import type { CoreTeamMember } from './coreTeamMember';

export interface ApplicationInput {
  startupName: string;
  websiteUrl?: string;
  founderNames: string;
  contactEmail: string;
  contactNumber?: string;
  linkedinFounder?: string;
  linkedinCompany?: string;
  headquarters?: string;
  yearOfIncorporation?: number;
  problemStatement?: string;
  solutionOverview?: string;
  sector: string;
  businessModel?: string;
  stage: string;
  coreProduct?: string;
  techStack?: string[];
  usp?: string;
  hasPatents?: boolean;
  patentDetails?: string;
  demoLink?: string;
  tam?: string;
  sam?: string;
  som?: string;
  customerSegment?: string;
  competitors?: string;
  competitiveAdvantage?: string;
  monthlyRevenue?: number;
  annualRevenue?: number;
  growthRate?: number;
  customerCount?: number;
  partnerships?: string;
  achievements?: string;
  fundingRaised?: number;
  previousInvestors?: string;
  burnRate?: number;
  runway?: number;
  projectionY1?: number;
  projectionY2?: number;
  projectionY3?: number;
  amountRaising?: number;
  currency?: string;
  fundingStage?: string;
  equityOffered?: number;
  useOfFunds?: string[];
  useOfFundsDetail?: string;
  founderEducation?: string;
  founderExperience?: string;
  teamSize?: number;
  coreTeam?: CoreTeamMember[];
  advisors?: string;
  whyFSV?: string;
  fsvValueAdd?: string;
  openToMentorship?: boolean;
  isRegistered?: boolean;
  hasLegalIssues?: boolean;
  legalExplanation?: string;
  consentGiven?: boolean;
}