# lib/api-zod/src/generated/types/applicationSubmitResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FSV Capital - Startup Funding Application API
 * OpenAPI spec version: 0.1.0
 */
import type { ScoreBreakdown } from './scoreBreakdown';

export interface ApplicationSubmitResult {
  success: boolean;
  applicationId: string;
  score: number;
  scoreBreakdown?: ScoreBreakdown;
}