# lib/api-zod/src/generated/types/coreTeamMember.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FSV Capital - Startup Funding Application API
 * OpenAPI spec version: 0.1.0
 */

export interface CoreTeamMember {
  name?: string;
  role?: string;
  linkedin?: string;
}