# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FSV Capital - Startup Funding Application API
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardStatsBySector } from './dashboardStatsBySector';
import type { DashboardStatsByStage } from './dashboardStatsByStage';
import type { DashboardStatsByStatus } from './dashboardStatsByStatus';

export interface DashboardStats {
  total: number;
  averageScore: number;
  shortlisted: number;
  thisMonth: number;
  bySector: DashboardStatsBySector;
  byStage: DashboardStatsByStage;
  byStatus: DashboardStatsByStatus;
}